Update README.md
The fairytale we chose to base our display on is Alice in Wonderland. The scene we chose to adapt into our project is the scene where Alice falls down the rabbit hole, and eats and drinks the cake and potion that change her size. In the scene we chose, Alice follows the rabbit which leads her to fall down the rabbit hole. When she gets to the bottom, she sees a table with cookies and a potion. She eats the cookies which cause her to grow and drinks the potion, which causes her to shrink. We chose this scene because there is a lot of interaction involved within it, which will help in thinking of interactions to integrate within the final project. Our goal for this project was to replicate the scene as accurately as possible with the time window we were given as well as incorporate five fully functional interactions.